Fix circular import part 2
