+++ src/datumaro/v2/__init__.py
@@ -2,8 +2,8 @@
 #
 # SPDX-License-Identifier: MIT
 
-from . import converters  # Import converters to register them
-from .converters import ConverterRegistry, converter, find_conversion_path
+from . import converters, find_conversion_path  # Import converters to register them
+from .converters import ConverterRegistry, converter
 from .dataset import Dataset, Sample
 from .export_import import export_dataset, import_dataset
 from .fields import Field, Semantic
